melissa4py/requirements.txt
pyzmq
mpi4py
tensorflow-gpu==1.13.2
keras
horovod